aisearch/requirements.txt
azure-search-documents==11.6.0b5
azure-identity==1.17.1
python-dotenv== 1.0.1
requests==2.32.3
